+++ CHANGELOG.md
@@ -10,6 +10,7 @@ New features:
 - CE-790 Add pattern matching filter to view top 200 scores per each site
 - CE-856 Add feature to export names/emails of all users
 - CE-753 Pattern Matching results are saved as species validations
+- Display the name of species on the spectrogram
 
 Resolved issues: